Removed redundant page
Removed a redundant listing page which was just used to link to the listing list page
