+++ src/components/App/Router/index.js
@@ -5,7 +5,6 @@ import Logout from "../../Auth/Logout";
 import Register from "../../Auth/Register";
 import Dashboard from "../../Dashboard";
 import Landing from "../../Landing";
-import Listing from "../../Listing";
 import Profile from "../../Profile";
 import Home from "../../Home"
 import { Role } from "../Authentication";
@@ -13,7 +12,7 @@ import ProtectedRoute from "../Authentication/ProtectedRoute";
 import ErrorRoute from "./ErrorRoute";
 import CreateListing from '../../Listing/CreateListing';
 import UpdateListing from '../../Listing/UpdateListing';
-import ListingList from '../../Listing/ListingList';
+import ListingList from '../../Listing';
 import ListingDisplay from '../../Listing/ListingDisplay';
 import Forgot from "../../Auth/Forgot";
 
@@ -28,7 +27,6 @@ const Router = () => (
     <ProtectedRoute exact roles={[Role.Admin]} path="/dashboard" component={Dashboard} />
     <ProtectedRoute exact roles={[Role.Flat]} path="/newlisting" component={CreateListing} />
     <ProtectedRoute exact path="/" component={Home} />
-    <ProtectedRoute exact roles={[Role.Flat]} path="/listing" component={Listing} />
     <ProtectedRoute exact roles={[Role.Flat]} path="/updatelisting" component={UpdateListing} />
     <ProtectedRoute exact roles={[Role.Flat]} path="/listings" component={ListingList} />
     <ProtectedRoute exact path="/listing/display" component={ListingDisplay} />
