+++ src/components/Listing/index.js
@@ -1,20 +1,19 @@
-import React, { useState } from 'react';
-import Avatar from '@material-ui/core/Avatar';
+import React, { useEffect, useState } from 'react';
+import { Link as RouterLink } from 'react-router-dom';
+import { Grid } from '@material-ui/core';
 import Button from '@material-ui/core/Button';
-// import ButtonGroup from '@material-ui/core/ButtonGroup';
+import ButtonGroup from '@material-ui/core/ButtonGroup';
+import axios from 'axios';
+import { useAuth } from '../App/Authentication';
+import Avatar from '@material-ui/core/Avatar';
 import CssBaseline from '@material-ui/core/CssBaseline';
-// import TextField from '@material-ui/core/TextField';
-// import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Link from '@material-ui/core/Link';
-// import Grid from '@material-ui/core/Grid';
 import Box from '@material-ui/core/Box';
 import LockOutlinedIcon from '@material-ui/icons/LockOutlined';
 import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import Container from '@material-ui/core/Container';
 import { Config } from '../../config';
-// import ListingRouter from './ListingRouter';
-// import { BrowserRouter } from 'react-router-dom';
 
 function Copyright() {
     return (
@@ -49,23 +48,52 @@ const useStyles = makeStyles((theme) => ({
     },
 }));
 
-export default function Listing(props) {
+//Shows the current user all listings they have created, along with the ooption to create a new listing
+function ListingList(props) {
     const classes = useStyles();
+    const { user, jwt } = useAuth();
+    const [listings, setListings] = useState([]);
 
-    //Decided to try using onSubmit buttons rather than routers, with a button state to control which route is taken
-
-    const [button, setButton] = useState(0);
-
-    const onSubmit = (e) => {
-        e.preventDefault();
-
-        if (button === 1) {
-            props.history.push('/listings');
-        }
-        if (button === 2) {
-            props.history.push('/newlisting');
-        }
+    //Passes the selected listing to the listing page for displaying
+    function selectListing(id) {
+        props.history.push({
+            pathname: '/listing/display',
+            state: { id: id },
+        });
     }
+
+    //Dynamically render individual buttons for each listing under the account
+    const renderButtons = () => {
+        let count = 0;
+        return listings.map((listing) => (
+            <Button
+                className="button"
+                variant="contained"
+                key={listing.id}
+                onClick={function () { selectListing(listing.id) }}
+            >
+                {++count}
+            </Button>
+        ))
+    }
+
+    //Fetch all listings owned by the current user on page load
+    useEffect(() => {
+        async function getListings() {
+            const URL = 'http://localhost:4000/listings/flat/'.concat(user.id);
+
+            const config = {
+                headers: { Authorization: `Bearer ${jwt}` }
+            };
+
+            const listings = await axios.get(URL, config);
+
+            setListings(listings.data);
+        }
+        // if (user.role === 'flat') {
+            getListings();
+        // }
+    }, [user, jwt])
 
     return (
         <Container component="main" maxWidth="xs">
@@ -77,32 +105,31 @@ export default function Listing(props) {
                 <Typography component="h1" variant="h5">
                     Listings
                 </Typography>
-                <form className={classes.form} onSubmit={onSubmit} noValidate>
-                    <Button
-                        onClick={() => (setButton(1))}
-                        type="submit"
-                        fullWidth
-                        variant="contained"
-                        color="primary"
-                        className={classes.submit}
-                    >
-                        Listings
-                    </Button>
-                    <Button
-                        onClick={() => (setButton(2))}
-                        type="submit"
-                        fullWidth
-                        variant="contained"
-                        color="primary"
-                        className={classes.submit}
-                    >
-                        Create Listing
-                    </Button>
-                </form>
+                    {renderButtons()}
+                    <br />
+                    <ButtonGroup variant="contained" color="primary">
+                        <Button
+                            className="button"
+                            component={RouterLink}
+                            to="/listings"
+                        >
+                            Account
+                        </Button>
+                        <Button
+                            className="button"
+                            component={RouterLink}
+                            to="/newlisting"
+                        >
+                            Create Listing
+                        </Button>
+                    </ButtonGroup>
+                {/* </Grid> */}
             </div>
             <Box mt={8}>
                 <Copyright />
             </Box>
         </Container>
     );
-};
+}
+
+export default ListingList;
